refactor(webui): extract ShadowRoot Shell and rename to renderer-surface
renderer-host 폴더를 renderer-surface로 rename한다. 본 commit에서
features/project/RenderedView.tsx의 re-export 경로도 ./renderer-surface/
로 갱신.

RendererLayer.tsx를 두 파일로 분해한다:
- ShadowShell.tsx: 컴포넌트 본체. ShadowShell 컴포넌트와 ShadowShellHandle
  타입을 새 이름으로 export.
- shadow-runtime.ts: LayerHostRuntime, layerHostRuntimes WeakMap,
  getLayerHostRuntime/disposeLayerHostRuntime, clearRendererStyles,
  injectCss, isRendererInstance helper.

기존 RendererLayer.tsx 파일은 임시 alias re-export shim으로 남긴다
(RendererLayer ↔ ShadowShell, RendererLayerHandle ↔ ShadowShellHandle).
호출처(RenderedView, useRendererHostMachine)의 import는 zero diff.
G1에서 호출처 갱신 + alias 제거 + DOM attr rename + dead code 정리.

issue #144 Phase F4.
