+++ apps/webui/src/client/features/project/RenderedView.tsx
@@ -1,1 +1,1 @@
-export { RenderedView } from "./renderer-host/RenderedView.js";
+export { RenderedView } from "./renderer-surface/RenderedView.js";
