+++ apps/webui/src/client/features/project/renderer-surface/RendererLayer.tsx
@@ -1,0 +1,3 @@
+// Temporary alias re-export. Removed in G1 once call sites use ShadowShell.
+export { ShadowShell as RendererLayer } from "./ShadowShell.js";
+export type { ShadowShellHandle as RendererLayerHandle } from "./ShadowShell.js";
